+++ index.html
@@ -7,7 +7,12 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>CAN 总线负载计算器</title>
     <link rel="stylesheet" href="styles.css" />
-    <script defer type="module" src="js/main.js"></script>
+    <link rel="modulepreload" href="main.js" />
+    <link rel="modulepreload" href="modules/ui/elements.js" />
+    <link rel="modulepreload" href="modules/ui/visualizer.js" />
+    <link rel="modulepreload" href="modules/calculator.js" />
+    <link rel="modulepreload" href="modules/config.js" />
+    <script defer type="module" src="main.js"></script>
 </head>
 
 <body>
@@ -20,9 +25,10 @@
             <div class="info-content">
                 <p>感谢 西南石油大学-铁人 制作的
                     <a href="https://bbs.robomaster.com/article/557395" target="_blank" title="RoboMaster 论坛相关文章">Excel
-                    表格</a>，此网页中的算法由 Claude
+                        表格</a>，此网页中的算法由 Claude
                     根据该表格编写，并由 <a href="https://lcwebsite.cn" target="_blank" title="LC的网站">LC</a>
-                    人工验证。欢迎为<a href="https://github.com/lc6464/CANbus-Load-Calculator" target="_blank" title="GitHub 存储库">此存储库</a>做出贡献！
+                    人工验证。欢迎为<a href="https://github.com/lc6464/CANbus-Load-Calculator" target="_blank"
+                        title="GitHub 存储库">此存储库</a>做出贡献！
                 </p>
                 <p>一般 CAN 总线负载在 80% 以内都是安全的，再高就要考虑各种干扰的影响了，有概率会开始丢包。</p>
                 <p>尽可能将多个不足 8 字节的包合并为一个数据帧，可以减少帧头帧尾的数量，以提高总线利用率。</p>
