--- js/ui/elements.js
@@ -1,13 +1,0 @@
-// js/ui/elements.js
-
-export const elements = {
-    baudRate: document.getElementById("baud-rate"),
-    canMessagesTable: document.getElementById("can-messages-table"),
-    addRowButton: document.getElementById("add-row-button"),
-    minLoadResult: document.getElementById("min-load-result"),
-    avgLoadResult: document.getElementById("avg-load-result"),
-    maxLoadResult: document.getElementById("max-load-result"),
-    loadBar: document.getElementById("load-bar"),
-    resetButton: document.getElementById("reset-button"),
-    visualizationContainer: document.getElementById("visualization-container")
-};
